% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_model.R, R/model_naive.R,
%   R/model_oscar.R, R/model_racer.R, R/model_roscar.R
\name{cate_model}
\alias{cate_model}
\alias{model_naive}
\alias{model_oscar}
\alias{model_racer}
\alias{model_roscar}
\title{Compute Model Coefficients}
\usage{
cate_model(
  rct,
  os = NULL,
  randomForest = FALSE,
  sampleSplit = FALSE,
  method = c("naive", "racer", "oscar", "roscar")
)

model_naive(rct, os = NULL, randomForest = FALSE)

model_oscar(rct, os)

model_racer(rct, os = NULL, randomForest = FALSE)

model_roscar(rct, os, sampleSplit = FALSE)
}
\arguments{
\item{rct}{Randomized clincal trial (RCT) data}

\item{os}{Observational study (OS) data}

\item{randomForest}{Utilize random forest approach}

\item{sampleSplit}{Utilize sampling to split data set}

\item{method}{By default, all methods are used. Can be
specified as \sQuote{naive}, \sQuote{racer}, \sQuote{oscar}, \sQuote{roscar}.}
}
\value{
\item{estimates by method}{
\describe{
\item{cate_coefficients}{cate estimate; not available in
naive/RACER random forest approach}
\item{cate_function}{function to predict CATE for new data}
\item{fitted_cate}{fitted CATE values}
\item{delta_coef_mat}{difference between coefficient estimates;
only available in OSCAR approach}
\item{cate_delta_coefficients}{difference between the two CATEs;
only available in R-OSCAR approach}
\item{cate_model}{CATE model; only available in naive random
forest approach}
}}
}
\description{
CATE estimates
}
\examples{
data(os, rct)
modNoOs <- cate_model(rct)
print(modNoOs, max = 40)
modOs <- cate_model(rct, os)
print(modOs, max = 40)

}
