% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cate.R
\name{estimate_cate}
\alias{estimate_cate}
\title{estimate_cate}
\usage{
estimate_cate(
  X,
  A,
  Y,
  propensity_vec,
  m_of_x_mat = NULL,
  normalize_weights = TRUE,
  mu_x_a_coefs = NULL,
  init_cate_coefs = NULL,
  K = 5,
  outcome_model_obs_coefs = NULL,
  useRF = FALSE
)
}
\arguments{
\item{X}{NEEDINFO}

\item{A}{NEEDINFO}

\item{Y}{NEEDINFO}

\item{propensity_vec}{NEEDINFO -- is this always `mean(A)`?}

\item{m_of_x_mat}{NEEDINFO}

\item{normalize_weights}{NEEDINFO}

\item{mu_x_a_coefs}{NEEDINFO (required for OSCAR approach)}

\item{init_cate_coefs}{NEEDINFO (required for R-OSCAR approach)}

\item{K}{NEEDINFO (required for sample split R-OSCAR approach)}

\item{outcome_model_obs_coefs}{NEEDINFO (required for sample split R-OSCAR approach)}

\item{useRF}{NEEDINFO (required for naive random forest approach)}
}
\value{
\item{cate_coefficients}{cate estimate; not available in
naive random forest approach}
\item{cate_function}{function to predict CATE for new data}
\item{fitted_cate}{fitted CATE values}
\item{delta_coef_mat}{difference between coefficient estimates;
only available in OSCAR approach}
\item{cate_delta_coefficients}{difference between the two CATEs;
only available in R-OSCAR approach}
\item{cate_model}{CATE model; only available in naive random
forest approach}
}
\description{
Estimate conditional average treatment effect.
}
\details{
Users will likely not call this function directly, instead
relying in behavior in \code{\link{model_coef}}.
}
\examples{
data(rct)
X <- rct$X;Y <- rct$Y;A <- rct$A
estimate_cate(X, A, Y, rep(mean(A), nrow(X)))$cate_coefficients[1:11,1]
}
