% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cate_oscar.R
\name{estimate_cate_oscar}
\alias{estimate_cate_oscar}
\title{estimate_cate}
\usage{
estimate_cate_oscar(X, A, Y, propensity_vec, mu_x_a_coefs)
}
\arguments{
\item{X}{NEEDINFO}

\item{A}{NEEDINFO}

\item{Y}{NEEDINFO}

\item{propensity_vec}{NEEDINFO}

\item{mu_x_a_coefs}{NEEDINFO}
}
\description{
Estimate conditional average treatment effect with R-OSCAR method under sample-split.
}
\keyword{internal}
