% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cate_r_oscar.R
\name{estimate_cate_r_oscar}
\alias{estimate_cate_r_oscar}
\title{estimate_cate}
\usage{
estimate_cate_r_oscar(
  X,
  A,
  Y,
  propensity_vec,
  init_cate_coefs,
  m_of_x_mat = NULL,
  normalize_weights = TRUE
)
}
\arguments{
\item{X}{NEEDINFO}

\item{A}{NEEDINFO}

\item{Y}{NEEDINFO}

\item{propensity_vec}{NEEDINFO}

\item{init_cate_coefs}{NEEDINFO}

\item{m_of_x_mat}{NEEDINFO}

\item{normalize_weights}{NEEDINFO}
}
\description{
Estimate conditional average treatment effect with R-OSCAR method.
}
\keyword{internal}
