% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cate_with_rct.R
\name{estimate_cate_with_rct}
\alias{estimate_cate_with_rct}
\title{estimate_cate}
\usage{
estimate_cate_with_rct(
  X,
  A,
  Y,
  propensity_vec,
  m_of_x_mat = NULL,
  normalize_weights = TRUE
)
}
\arguments{
\item{X}{NEEDINFO}

\item{A}{NEEDINFO}

\item{Y}{NEEDINFO}

\item{propensity_vec}{NEEDINFO}

\item{m_of_x_mat}{NEEDINFO}

\item{normalize_weights}{NEEDINFO}
}
\description{
Estimate conditional average treatment effect with naive/RACER approach.
}
\keyword{internal}
