% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_outcome_model_by_arm.R
\name{fit_outcome_model_by_arm}
\alias{fit_outcome_model_by_arm}
\title{Fit Outcome Model}
\usage{
fit_outcome_model_by_arm(X, A, Y)
}
\arguments{
\item{X}{NEEDINFO}

\item{A}{NEEDINFO}

\item{Y}{NEEDINFO}
}
\description{
Fit outcome regression model for each treatment arm.
}
\keyword{internal}
