% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eval.R
\name{model_eval}
\alias{model_eval}
\title{Model Evaluation}
\usage{
model_eval(mod, tau)
}
\arguments{
\item{mod}{\sQuote{cate_model} object, created with \code{\link{cate_model}}.}

\item{tau}{True treatment effect for RCT data.}
}
\value{
\item{rmse}{Root mean squared error}
\item{rank_corr}{Spearman correlation}
\item{accuracy_itr}{Model accuracy with respect to sign}
\item{cal_intercept}{Calibration intercept}
\item{cal_slope}{Calibration coefficients}
}
\description{
Compare model estimates
}
\examples{
data(os, rct, tau)
mod <- cate_model(rct, os)
model_eval(mod, tau)

}
