% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_naive.R
\name{model_naive}
\alias{model_naive}
\title{model_naive}
\usage{
model_naive(rct, os = NULL, randomForest = FALSE)
}
\description{
Estimate conditional average treatment effect.
}
\details{
Users will likely not call this function directly, instead
relying in behavior in \code{\link{cate_model}}.
}
