% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_oscar.R
\name{model_oscar}
\alias{model_oscar}
\title{model_oscar}
\usage{
model_oscar(rct, os)
}
\description{
Estimate conditional average treatment effect.
}
\details{
Users will likely not call this function directly, instead
relying in behavior in \code{\link{cate_model}}.
}
