% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_racer.R
\name{model_racer}
\alias{model_racer}
\title{model_racer}
\usage{
model_racer(rct, os = NULL, randomForest = FALSE)
}
\description{
Estimate conditional average treatment effect.
}
\details{
Users will likely not call this function directly, instead
relying in behavior in \code{\link{cate_model}}.
}
