% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_roscar.R
\name{model_roscar}
\alias{model_roscar}
\title{model_roscar}
\usage{
model_roscar(rct, os, sampleSplit = FALSE)
}
\description{
Estimate conditional average treatment effect.
}
\details{
Users will likely not call this function directly, instead
relying in behavior in \code{\link{cate_model}}.
}
