% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\docType{data}
\name{os}
\alias{os}
\title{Example Observational Study Data}
\format{
A list with three components.
\describe{
  \item{X}{A numeric matrix, with 10,000 rows and 100 columns; independent variables}
  \item{Y}{A numeric vector of length 10,000; dependent variable}
  \item{A}{A numeric vector of length 10,000; indicates treatment group}
}
}
\usage{
data(os)
}
\description{
Randomized data for an observational study.
}
\examples{
data(os)
}
\keyword{datasets}
