% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cate_model}
\alias{print.cate_model}
\title{Print methods}
\usage{
\method{print}{cate_model}(x, ...)
}
\arguments{
\item{x}{A cate_model object.}

\item{\dots}{Additional parameters passed to print.data.frame.}
}
\description{
Print method for \sQuote{cate_model} objects.
}
\details{
Print its argument and return it invisibly.
}
\examples{
data(os, rct)
mod <- cate_model(rct, os)
print(mod, max = 40)

}
