% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct.R
\docType{data}
\name{rct}
\alias{rct}
\title{Example Randomized Clincal Trial Data}
\format{
A list with three components.
\describe{
  \item{X}{A numeric matrix, with 1,000 rows and 100 columns; independent variables}
  \item{Y}{A numeric vector of length 1,000; dependent variable}
  \item{A}{A numeric vector of length 1,000; indicates treatment group}
}
}
\usage{
data(rct)
}
\description{
Randomized data for a randomized clinical trial.
}
\examples{
data(rct)
}
\keyword{datasets}
