% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSCAR-package.R
\docType{package}
\name{roscar-package}
\alias{roscar}
\alias{roscar-package}
\title{Package}
\description{
R-OSCAR (Robust Observational Studies for CMO-Augmented RCT)
is a two-stage estimator that calibrates OS outcome predictions
to the RCT population and corrects residual biases through
regularized regression. Comparison methods are also provided.
}
\examples{
data(os, rct, tau)
mod <- cate_model(rct, os, randomForest = TRUE)
head(model_pred(mod))
head(model_pred(mod, rct$X))
model_eval(mod, tau)
}
\author{
Amir Asiaeetaheri \email{amir.asiaeetaheri@vumc.org}

Maintainer: Amir Asiaeetaheri \email{amir.asiaeetaheri@vumc.org}
}
