% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{simulate_rct_and_os_data}
\alias{simulate_rct_and_os_data}
\title{Simulate RCT and OS Data}
\usage{
simulate_rct_and_os_data(
  n_r = 1000,
  n_o = 10000,
  p = 100,
  support_fraction = 1/10,
  covariate_effect = 2/3,
  delta_fraction = 2/100,
  treatment_prop = 1/3,
  shift_x = TRUE,
  shift_y = TRUE,
  model = c("linear", "quadratic", "sinusoidal"),
  percent_modifier_to_cover = NULL,
  frac_U = NULL,
  frac_V = NULL
)
}
\arguments{
\item{n_r}{number of observations in RCT data}

\item{n_o}{number of observations in OS data}

\item{p}{total number of covariates to create}

\item{support_fraction}{percent of covariates associated with observed betas}

\item{covariate_effect}{error in the outcome model. In the covariate mismatch
setting this should be a named vector with values for \sQuote{Z} and \sQuote{V}.}

\item{delta_fraction}{percent of betas to shift for RCT}

\item{treatment_prop}{percent of OS to receive treatment}

\item{shift_x}{set to TRUE to introduce covariate shift}

\item{shift_y}{set to TRUE to introduce outcome shift}

\item{model}{function to model outcome; defaults to a linear model. Can
also be quadratic, sine, or user specified.}

\item{percent_modifier_to_cover}{when greater than zero, use missing
effect modifier setting; incompatible with covariate mismatch setting}

\item{frac_U}{fraction of RCT data not found in OS data}

\item{frac_V}{fraction of OS data not found in RCT data}
}
\value{
\item{simulated data, components and settings}{
\describe{
\item{X_RCT}{RCT covariates}
\item{X_OS}{OS covariates}
\item{components}{components created in covariate mismatch settting}
\item{a_r}{treatment assignment for RCT data}
\item{y_r}{0/1 outcome for RCT data}
\item{a_o}{treatment assignment for OS data}
\item{y_o}{0/1 outcome for OS data}
\item{beta_RCT}{betas for RCT data}
\item{beta_OS}{betas for OS data}
\item{idx}{indices used to track U/Z/V in covariate mismatch settting}
\item{exp_settings}{preserved function argument values}
}}
}
\description{
Simulate randomized clinical trial and observational data
with shared properties.
}
\examples{
x1 <- simulate_rct_and_os_data(
  n_r = 300, n_o = 1000, p = 100,
)
# simulate outcome with quadratic function
x2 <- simulate_rct_and_os_data(
  n_r = 300, n_o = 1000, model = 'quadratic'
)
# simulate missing effect modifiers
x3 <- simulate_rct_and_os_data(
  n_r = 300, n_o = 1000, percent_modifier_to_cover = 0.2
)
# simulate with covariate mismatch
x4 <- simulate_rct_and_os_data(
  n_r = 300, n_o = 1000, support_fraction = 1/20,
  covariate_effect = c(Z = 2/3, V = 1),
  frac_U = 0.3, frac_V = 0.3
)

}
