% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_estimate_cate_r_oscar.R
\name{ss_estimate_cate_r_oscar}
\alias{ss_estimate_cate_r_oscar}
\title{estimate_cate}
\usage{
ss_estimate_cate_r_oscar(
  X,
  A,
  Y,
  propensity_vec,
  K = 5,
  outcome_model_obs_coefs,
  m_of_x_mat = NULL,
  normalize_weights = TRUE
)
}
\arguments{
\item{X}{NEEDINFO}

\item{A}{NEEDINFO}

\item{Y}{NEEDINFO}

\item{propensity_vec}{NEEDINFO}

\item{K}{NEEDINFO}

\item{outcome_model_obs_coefs}{NEEDINFO}

\item{m_of_x_mat}{NEEDINFO}

\item{normalize_weights}{NEEDINFO}
}
\description{
Estimate conditional average treatment effect with R-OSCAR method under sample-split.
}
\keyword{internal}
