% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\docType{data}
\name{tau}
\alias{tau}
\title{True Treatment Effect}
\format{
A numeric vector of length 1,000.
}
\usage{
data(tau)
}
\description{
True treatment effect from RCT data.
}
\examples{
data(tau)
}
\keyword{datasets}
