% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAint-package.R
\docType{package}
\name{AAint-package}
\alias{AAint}
\alias{AAint-package}
\title{Package}
\description{
what does it do?
}
\examples{
data(os, rct, tau)
mod <- cate_model(rct, os, randomForest = TRUE)
head(model_pred(mod))
head(model_pred(mod, rct$X))
model_eval(mod, tau)
}
\author{
Amir Asiaeetaheri \email{amir.asiaeetaheri@vumc.org}

Maintainer: Amir Asiaeetaheri \email{amir.asiaeetaheri@vumc.org}
}
