% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data.R
\name{build_data}
\alias{build_data}
\alias{build_data.default}
\alias{build_data.sim_data}
\alias{impute_OLS}
\alias{impute_0}
\title{Build RCT and OS}
\usage{
build_data(RCT, OS, RCT_imp_method, OS_imp_method)

\method{build_data}{default}(RCT, OS, RCT_imp_method = impute_OLS, OS_imp_method = impute_0)

\method{build_data}{sim_data}(dat, RCT_imp_method = impute_OLS, OS_imp_method = impute_0)

impute_OLS(has_excl, has_shr, oth_shr, oth_excl)

impute_0(has_excl, has_shr, oth_shr, oth_excl)
}
\arguments{
\item{RCT}{Randomized clincal trial (RCT) data
is a list with U (covariates not available in OS),
Z (shared covariates with OS), Y (outcome), A (treatment).}

\item{OS}{Observational study (OS) data is a list
with V (covariates not available in RCT),
Z (shared covariates with RCT), Y, A.}

\item{RCT_imp_method}{Imputation function for RCT data;
when unspecified \sQuote{Vhat_r} will be imputed with \sQuote{OLS}.
Set to NULL to only use shared covariates (no imputation).}

\item{OS_imp_method}{Imputation function for OS data; when
unspecified \sQuote{Uhat_o} will be imputed as zeroes.
Set to NULL to only use shared covariates (no imputation).}

\item{dat}{Instead of \sQuote{RCT} and \sQuote{OS}, a \sQuote{sim_data}
object can be provided (see \code{\link{simulate_rct_and_os_data}}).}

\item{has_excl}{Required for imputation method}

\item{has_shr}{Required for imputation method}

\item{oth_shr}{Required for imputation method}

\item{oth_excl}{Required for imputation method

Imputation functions can be provided to impute either with RCT or
OS data. In either case, it's expected that the function will have four
arguments (has_excl, has_shr, oth_shr, oth_excl). \sQuote{has_excl}
and \sQuote{has_shr} refer to the data set receiving imputation.
\sQuote{oth_shr} and \sQuote{oth_excl} refer to the data set being
used for imputation. In the typical setting of imputing RCT covariates
from OS, has_excl = RCT[['U']], has_shr = RCT[['Z']], oth_shr = OS[['Z']],
and oth_excl = OS[['V']]. A user-defined function is limited to this
information.}
}
\value{
\item{components of RCT and OS data}{
\describe{
\item{RCT}{covariates (X), outcome (Y), and treatment (A) for RCT data}
\item{OS}{covariates (X), outcome (Y), and treatment (A) for OS data}
\item{tau}{True treatment effect for RCT data, only available
with \sQuote{sim_data} method.}
}}
}
\description{
Create necessary RCT and OS components required to
run \code{\link{cate_model}}.
}
\examples{
data(os, rct)
rct1 <- list(Z = rct$X, Y = rct$Y, A = rct$A)
os1 <- list(Z = os$X, Y = os$Y, A = os$A)
ex1 <- build_data(rct1, os1)

# simulate with covariate mismatch
x <- simulate_rct_and_os_data(
  n_r = 300, n_o = 1000, support_fraction = 1/20,
  covariate_effect = c(Z = 2/3, V = 1),
  frac_U = 0.3, frac_V = 0.3
)
rct_os <- build_data(x)

}
