% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_outcome_model.R
\name{calibrate_outcome_model}
\alias{calibrate_outcome_model}
\title{Calibrate Outcome Model}
\usage{
calibrate_outcome_model(X, A, Y, mu_x_a_coefs)
}
\arguments{
\item{X}{Data set covariates}

\item{A}{Vector with treatment assignment}

\item{Y}{Vector with outcome variable}

\item{mu_x_a_coefs}{Outcome regression coefficient matrix}
}
\description{
Calibrate the outcome regression functions to (RCT) data.
}
\keyword{internal}
