% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_model.R
\name{cate_model}
\alias{cate_model}
\title{Compute Model Coefficients}
\usage{
cate_model(
  rct,
  os = NULL,
  randomForest = FALSE,
  sampleSplit = FALSE,
  method = c("naive", "racer", "oscar", "roscar")
)
}
\arguments{
\item{rct}{NEEDINFO}

\item{os}{NEEDINFO}

\item{randomForest}{NEEDINFO}

\item{sampleSplit}{NEEDINFO}
}
\value{
\item{estimates by method}{
\describe{
\item{cate_coefficients}{cate estimate; not available in
naive/RACER random forest approach}
\item{cate_function}{function to predict CATE for new data}
\item{fitted_cate}{fitted CATE values}
\item{delta_coef_mat}{difference between coefficient estimates;
only available in OSCAR approach}
\item{cate_delta_coefficients}{difference between the two CATEs;
only available in R-OSCAR approach}
\item{cate_model}{CATE model; only available in naive random
forest approach}
}}
}
\description{
CATE estimates
}
\examples{
data(os, rct)
modNoOs <- cate_model(rct)
print(modNoOs, max = 40)
modOs <- cate_model(rct, os)
print(modOs, max = 40)

}
