% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_outcome_model_by_arm_rf.R
\name{fit_outcome_model_by_arm_rf}
\alias{fit_outcome_model_by_arm_rf}
\title{Fit Outcome Model with Random Forest}
\usage{
fit_outcome_model_by_arm_rf(X, A, Y)
}
\arguments{
\item{X}{Data set covariates}

\item{A}{Vector with treatment assignment}

\item{Y}{Vector with outcome variable}
}
\description{
Fit outcome regression model for each treatment arm.
}
\keyword{internal}
