% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eval.R
\name{model_eval}
\alias{model_eval}
\title{Model Evaluation}
\usage{
model_eval(mod, tau)
}
\arguments{
\item{mod}{\sQuote{cate_model} object, created with \code{\link{cate_model}}.}

\item{tau}{NEEDINFO}
}
\value{
\item{rmse}{NEEDINFO}
\item{rank_corr}{NEEDINFO}
\item{accuracy_itr}{NEEDINFO}
\item{cal_intercept}{NEEDINFO}
\item{cal_slope}{NEEDINFO}
}
\description{
Compare model estimates
}
\examples{
data(os, rct, tau)
mod <- cate_model(rct, os)
model_eval(mod, tau)

}
