% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pred.R
\name{model_pred}
\alias{model_pred}
\title{Model Predict}
\usage{
model_pred(object, newdata)
}
\arguments{
\item{object}{\sQuote{cate_model} object, created with \code{\link{model_coef}}.}

\item{newdata}{An optional data frame in which to look for
variables with which to predict. If omitted, the fitted values are used.}
}
\value{
A data.frame with estimates from each method.
}
\description{
CATE estimates with new data
}
\examples{
data(os, rct)
mod <- model_coef(rct, os)
prd <- model_pred(mod)
prdNewdata <- model_pred(mod, rct$X)

}
